lesson_4: add posts service and query with commands repo
