+++ lesson_4/src/routes/testing.routes.ts
@@ -1,7 +1,7 @@
-import { type Express, type Response, type Request, Router } from "express";
+import { type Response, type Request, Router } from "express";
 import { ENDPOINTS, HTTP_STATUS } from "../constants";
 import { blogsCommandsRepository } from "../repositories/blogs.commands.repository";
-import { postsRepository } from "../repositories/posts.repository";
+import { postsCommandsRepository } from "../repositories/posts.commands.repository";
 
 export const testingRouter = Router({});
 
@@ -9,7 +9,7 @@ testingRouter
   .route(ENDPOINTS.TESTING)
   .delete((_req: Request, res: Response) => {
     blogsCommandsRepository.clearBlogs();
-    postsRepository.clearPosts();
+    postsCommandsRepository.clearPosts();
 
     res.sendStatus(HTTP_STATUS.NO_CONTENT);
   });
